SeparateMiddlewareMethods:  Completed the separation of middleware methods from the controller methods, and are being imported from their own middleware namespace.
