+++ app/controllers/AuthController/index.js
@@ -1,6 +1,5 @@
 import User from '../../models/user'
 import jwt from 'jsonwebtoken'; // to generate signed token
-import expressJwt from 'express-jwt'; // authorization check
 import { errorHandler } from '../../helpers/dbErrorHandler'
 
 export const register = (req, res) => {
@@ -70,36 +69,4 @@ export const logout = (req, res) => {
     res.json({
         message: 'Logout is successful'
     })
-}
-
-export const requireLogin = expressJwt({
-    secret: process.env.JWT_SECRET,
-    userProperty: 'auth'
-})
-
-export const isAuth = (req, res, next) => {
-         
-    console.log('user profile: ', req.profile, 'req.auth: ', req.auth)
-    let user = req.profile && req.auth && (req.profile._id == req.auth._id)
-    console.log('user: ', user)
-
-   if (!user) {
-       return res.status(403).json({
-           error: 'Access denied!'
-       })
-   }
-
-   next()
-}
-
-export const isAdmin = (req, res, next) => {
-
-   // role = 0 is user, role = 1 is admin
-   if (req.profile.role === 0) {
-       return res.status(403).json({
-           error: 'Admin resource! Access denied'
-       })
-   }
-
-   next()
 }
