+++ app/controllers/UserController/index.js
@@ -1,15 +1,15 @@
 import User from '../../models/User'
 
-export const userById = (req, res, next, id) => {
-	User.findById(id).exec((err, user) => {
-		if (err || !user) {
-			return res.status(400).json({
-				error: 'User was not found'
-			})
-		}
+// export const userById = (req, res, next, id) => {
+// 	User.findById(id).exec((err, user) => {
+// 		if (err || !user) {
+// 			return res.status(400).json({
+// 				error: 'User was not found'
+// 			})
+// 		}
 
-		req.profile = user
-		next()
+// 		req.profile = user
+// 		next()
 
-	})
-}
+// 	})
+// }
